Move HREF type into its own file, rename GetURL
